quinta subida proyecto ajax implementado

+++ app/Http/Controllers/work_areaController.php
@@ -35,7 +35,10 @@ class work_areaController extends Controller
     {
         if($request->ajax())
         {
-        $work_area = work_area::find($id)->update($request->all());
+        $work_area = work_area::find($id);
+        $work_area->name = $request->get('name');
+        $work_area->description = $request->get('description');
+        $work_area->update();
         return response(['success' => true, 'message' => 'Area de trabajo actualizada correctamente, id:' . $work_area->id, 'data' => $work_area], 201)
                     ->header('Content-Type', 'text/plain');
         }
@@ -44,8 +47,15 @@ class work_areaController extends Controller
 
     public function destroy($id)
     {
+        $work_area = work_area::find($id);
+        if (isset($work_area)) {
         work_area::find($id)->delete();
-
-        return response()->json(['done']);
+        return response(['success' => true, 'message' => 'Eliminado', 'data' => $work_area->id], 201)
+                    ->header('Content-Type', 'text/plain');
+        }
+        else{
+            return response(['success' => false, 'message' => 'Registro no se puede eliminar porque no existe', 'data' => $work_area->id], 400)
+                ->header('Content-Type', 'text/plain');
+        }
     }
 }
